Refactor input and update README
