+++ src/lib.rs
@@ -2,26 +2,30 @@ extern crate pest;
 extern crate pest_derive;
 use std::fs;
 use std::error::Error;
-mod day1;
+mod day01;
 
 pub fn run(config: Config) -> Result<String, Box<dyn Error>> {
-    match config.command.as_str() {
-        "day1a" => day1::run_a(&config.file_path),
-        "day1b" => day1::run_a(&config.file_path),
-        _ => Err(format!("Unrecognized command '{}'", &config.command))?
+    match config.day.as_str() {
+        "01a" => day01::run_a(&config.file_path),
+        "01b" => day01::run_a(&config.file_path),
+        _ => Err(format!("Unrecognized command '{}'", &config.day))?
     }
 }
 pub struct Config {
-    pub command: String,
+    pub day: String,
     pub file_path: String,
     pub args: Vec<String>,
 }
 
 impl Config {
     pub fn new(args: Vec<String>) -> Config {
+        let ex = match args[2].as_str() {
+            "ex" => "_ex",
+            _ => "",
+        };
         Config {
-            command: args[1].clone(),
-            file_path: args[2].clone(),
+            day: args[1].clone(),
+            file_path: format!("data/{}{}.txt", args[1], ex),
             args: args.into_iter().skip(3).collect(),
         }
     }
